MonsterMesh: 386 daycare social + Gen2/3 boarding + battle XP + 8-skin sprites
- Daycare: 386-species social vignettes with partner-perspective remoteMessage;
  social prioritized when neighbors present; Gen 2/3 daycare boarding
  (speciesDex widened to uint16).
- Battles write result-based XP back to the Game Boy save (D5).
- Sprites re-baked as 4bpp indexed + normal/shiny palettes (Gen3Front/Back4bpp.h),
  ~82KB smaller than the RGB565 set.
- 8 colour skins on the battle screen ('V' cycles): Regular, Shiny, Rainbow
  (light eyes, animated), Dark (red-balloon splash, gamma 2.1), Dark-Shiny,
  Pink, Dark-Pink, Dark-Rainbow (animated). uint16 sprite species cache fix.

+++ src/modules/monstermesh/DaycareTypes.h
@@ -247,7 +247,7 @@ inline const char *rivalryTierName(uint8_t rivalry) {
 // ── Per-Pokemon daycare state ────────────────────────────────────────────────
 
 struct DaycarePokemonState {
-    uint8_t  speciesDex;          // Pokedex number (1-151)
+    uint16_t speciesDex;         // national dex (1-386)
     char     nickname[11];        // Player-set nickname (10 + null), empty = use species name
     uint8_t  savLevel;            // Level read from SAV file at check-in
     uint32_t savExp;              // Total EXP read from SAV file at check-in
